requirements.txt
django==3.2.10
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
psycopg2-binary==2.9.2
PyJWT==2.3.0